fix&feat: Code crystalization ans swagger docs
