+++ src/database/zorm.service.ts
@@ -1,4 +1,5 @@
 import { randomUUID } from 'crypto';
+import { User } from 'src/models/user.interface';
 import { DatabaseService } from './database.service';
 
 // HandMade Custom Orm
@@ -60,5 +61,51 @@ export class CustomOrm {
     const pattern =
       /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/;
     return pattern.test(str);
+  }
+
+  // UserSpecific Layer
+  createUser(UserData): any {
+    const { login, password } = UserData;
+    const match = this.databaseService.findMatching('users', 'login', login);
+
+    if (match) {
+      return { success: false, error: 'Login already taken!' };
+    }
+
+    const uuid = randomUUID();
+    const userData = {
+      id: uuid,
+      login,
+      password,
+      version: 1,
+      createdAt: Date.now(),
+      updatedAt: Date.now(),
+    };
+
+    this.databaseService.insert('users', userData);
+
+    return this.deepClone({ success: false, data: userData });
+  }
+
+  updateUser(user: User): boolean {
+    try {
+      let users = this.databaseService.get('users');
+
+      user.updatedAt = Date.now();
+      user.version = user.version + 1;
+
+      users = users.map((usr: User) =>
+        usr.id === user.id ? { ...usr, ...user } : usr,
+      );
+      this.databaseService.set('users', users);
+      return true;
+    } catch (error) {
+      console.log('Error: ', error);
+      return false;
+    }
+  }
+
+  deepClone(data) {
+    return data ? JSON.parse(JSON.stringify(data)) : null;
   }
 }
